Move TestParseUriParams to test_image_provider and drop duplicate URI regex tests
TestParseUriParams tests a general kivy.core.image utility (_parse_uri_params),
not anything SVG-specific, so it belongs in test_image_provider.py. TestProviderUriRegex
is removed from test_thorvg_svg_provider.py as it is fully covered by the existing
TestImageProviderURIScheme class in test_image_provider.py.
