+++ kivy/tests/test_thorvg_svg_provider.py
@@ -78,104 +78,6 @@ class TestApplyMinRasterSize:
 
 
 # ---------------------------------------------------------------------------
-# _parse_uri_params
-# ---------------------------------------------------------------------------
-
-class TestParseUriParams:
-    """Pure unit tests for the URI parameter parser."""
-
-    @pytest.fixture(autouse=True)
-    def import_helper(self):
-        from kivy.core.image import _parse_uri_params
-        self.fn = _parse_uri_params
-
-    def test_none_returns_empty(self):
-        assert self.fn(None) == {}
-
-    def test_empty_string_returns_empty(self):
-        assert self.fn('') == {}
-
-    def test_single_int_param(self):
-        assert self.fn('size=512') == {'size': 512}
-
-    def test_multiple_params(self):
-        result = self.fn('size=256,quality=high')
-        assert result['size'] == 256
-        assert result['quality'] == 'high'
-
-    def test_int_conversion(self):
-        result = self.fn('size=128')
-        assert isinstance(result['size'], int)
-
-    def test_non_int_stays_string(self):
-        result = self.fn('mode=contain')
-        assert result['mode'] == 'contain'
-
-    def test_whitespace_around_tokens_ignored(self):
-        result = self.fn(' size = 512 ')
-        assert result['size'] == 512
-
-    def test_token_without_equals_ignored(self):
-        result = self.fn('badtoken,size=64')
-        assert 'badtoken' not in result
-        assert result['size'] == 64
-
-
-# ---------------------------------------------------------------------------
-# URI regex
-# ---------------------------------------------------------------------------
-
-class TestProviderUriRegex:
-    """Validate the updated _provider_uri_re with optional [params] block."""
-
-    @pytest.fixture(autouse=True)
-    def import_re(self):
-        from kivy.core.image import _provider_uri_re
-        self.re = _provider_uri_re
-
-    def test_basic_uri_no_params(self):
-        m = self.re.match('@image_provider:pil(image.png)')
-        assert m is not None
-        assert m.group(1) == 'pil'
-        assert m.group(2) is None
-        assert m.group(3) == 'image.png'
-
-    def test_uri_with_size_param(self):
-        m = self.re.match('@image_provider:thorvg_svg[size=256](icon.svg)')
-        assert m is not None
-        assert m.group(1) == 'thorvg_svg'
-        assert m.group(2) == 'size=256'
-        assert m.group(3) == 'icon.svg'
-
-    def test_uri_with_multiple_params(self):
-        m = self.re.match('@image_provider:thorvg_svg[size=512,mode=fit](a.svg)')
-        assert m is not None
-        assert m.group(2) == 'size=512,mode=fit'
-
-    def test_uri_with_absolute_path(self):
-        m = self.re.match('@image_provider:sdl3(/abs/path/image.jpg)')
-        assert m is not None
-        assert m.group(3) == '/abs/path/image.jpg'
-
-    def test_uri_with_relative_path(self):
-        m = self.re.match('@image_provider:imageio(../imgs/icon.png)')
-        assert m is not None
-        assert m.group(3) == '../imgs/icon.png'
-
-    def test_plain_path_does_not_match(self):
-        assert self.re.match('image.png') is None
-
-    def test_wrong_prefix_does_not_match(self):
-        assert self.re.match('@provider:pil(image.png)') is None
-
-    def test_missing_provider_name_does_not_match(self):
-        assert self.re.match('@image_provider:(image.png)') is None
-
-    def test_missing_path_does_not_match(self):
-        assert self.re.match('@image_provider:pil') is None
-
-
-# ---------------------------------------------------------------------------
 # Config default
 # ---------------------------------------------------------------------------
